test(deepseek-v41): every compared run takes its roundings and picks from its own float64 twin
In the pytest run on the Colab L4 VM, which ran on the host CPU because
tests/conftest.py defaults JAX_PLATFORMS to cpu and the run set none, the
quantization-aware cached prefill rounded one window key the other way
from the reference. The key sits in the last layer's window at row 1,
position 1, channel 28; its input is 1.6e-7 of its block's amax from an
FP8 rounding boundary, the seed rank's closest window element at 0.04 of
NOISE. That one step put qat_prompt_logits 1.505e-6 from the float64
truth, 2.56 times the reference's 5.88e-7, where tests/reference_error.py
allows 2 (test_the_quantization_aware_cache_and_drafter_match_the_reference
at b1910240). The box CPU does not round it that way, and neither does the
L4 GPU (6.9e-7, ratio 1.18). Of the 119 seeds the rank scored, none keeps
every element clear of NOISE (the best leaves one at 0.12 of it), so no
seed rule holds the RMS rule across machines.

decided(run, model, variables) in tools/deepseek_v41_numerics.py runs
`run` twice. Its float64 twin runs first: x64 on, the model, its
variables and every fp32 pin Dew names as jnp.float32 in float64, each
quantizer rounding its input's float32, which rounds as the float64
value does but at a tie; ordered callbacks record each quantizer output
and top-k pick. The given fp32 run follows, each quantizer and top-k
taking the twin's decision back through an ordered io_callback, and
recording the quantizer inputs and top-k rows it meets. Every comparison
with the reference (both forwards, both losses and updates, both cached
runs with their drafts, the vision half) and the input check go through
it, so a value within fp32 noise of a boundary goes the way the exact
value goes. That is the reference's way: `--fp64` now refuses a fixture
whose recorded roundings, picks or drafts its float64 run makes
otherwise (none at SEED 106). The quantizers' own rounding stays held
bit for bit to the kernels and to every recorded call, compiled on the
device. captured and recorded_outputs go; straight_through is public for
the replay.

Each test holds the twin too (twin_close): its RMS distance from the
truth at most FACTOR times the reference's own fp32 rounding scaled by
float64's finer unit, eps64 / eps32 = 2**-29, and the loss by loss_close
alike. A rounding or pick the twin took otherwise than the truth would
move an output by fp32-sized steps, far past that, so the tests assert
directly that the twin decides as the truth does (DeepSeekV41Review's
suggestion on N2). The twin of the export test's forward starts from the
fp32 step, which the truth does not take, so only its decisions count.

A media model's twin asks its towers for HIGHEST precision, at which
float64 multiplies anyway: jax's own XLA attention takes its softmax in
float32 whatever its inputs, which left the vision twin 2e-8 to 1e-7 from
the truth, and HIGHEST sends the ViT's attention down Dew's reference
path, whose fp32 pins widen.

The fp64 mode folds into residuals, which reports each twin's distance
from the truth: 7e-16 to 5e-14 for every output, the vision half's
included (2.3e-16, 2.5e-15, 1.7e-15), so Dew's arithmetic is the
reference's. vision_run compiles its calls. Two scatters write their
int32 target's dtype, which the twin's int64 positions and tokens
otherwise promote (a jax FutureWarning).

Box CPU on this commit: 21 passed. residuals on CPU: every ratio 0.50 to
1.35 but the loss's, which loss_close bounds (11.97 is the reference's
loss sitting 1.5e-7 from the truth by chance, below its fp32 spacing of
4.8e-7); qat_prompt_logits 1.08. The twins sit 0.43 to 1.5 times the
scaled rounding (decode_logits the largest).

+++ src/dew/nn/dspark.py
@@ -156,7 +156,7 @@ def draft(stages, spec: DSpark, embed: Callable, logits_of: Callable, hc_mult: i
             stage.seed(store)
         return None
     block = jnp.full((tokens.shape[0], spec.block_size), spec.noise_token_id, jnp.int32)
-    block = block.at[:, 0].set(tokens)
+    block = block.at[:, 0].set(tokens.astype(block.dtype))
     streams = expand_streams(embed(block), hc_mult)
     carry = Carried(streams, first_stream(streams))
     for stage in stages:
